chore(ai): migrate createContent + listContent to shared tool definitions
